fix: resolve blank page crash - Decimal to float serialization, safe Number() wrapping, route encoding fix
Made-with: Cursor

+++ Backend/app/schemas/counter_offer_schema.py
@@ -1,20 +1,19 @@
 from datetime import datetime
-from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, Field
 
 
 class CounterOfferCreate(BaseModel):
-    counter_price: Decimal = Field(..., gt=0, max_digits=12, decimal_places=2)
+    counter_price: float = Field(..., gt=0)
     message: str | None = None
 
 
 class CounterOfferRead(BaseModel):
     id: UUID
     request_id: UUID
-    original_price: Decimal
-    counter_price: Decimal
+    original_price: float
+    counter_price: float
     message: str | None = None
     created_by: UUID
     creator_name: str | None = None
